+++ azure-functions-maven-plugin/src/main/java/com/microsoft/azure/maven/function/DeployMojo.java
@@ -17,7 +17,6 @@ import com.microsoft.azure.toolkit.lib.appservice.function.AzureFunctions;
 import com.microsoft.azure.toolkit.lib.appservice.function.FunctionApp;
 import com.microsoft.azure.toolkit.lib.appservice.function.FunctionAppBase;
 import com.microsoft.azure.toolkit.lib.appservice.function.FunctionsServiceSubscription;
-import com.microsoft.azure.toolkit.lib.appservice.model.FunctionAppRuntime;
 import com.microsoft.azure.toolkit.lib.appservice.model.FunctionDeployType;
 import com.microsoft.azure.toolkit.lib.appservice.model.OperatingSystem;
 import com.microsoft.azure.toolkit.lib.appservice.model.PricingTier;
@@ -138,6 +137,7 @@ public class DeployMojo extends AbstractFunctionMojo {
         validateFunctionCompatibility();
         validateArtifactCompileVersion();
         validateApplicationInsightsConfiguration();
+        // validate container apps hosting of function app
     }
 
     private void validateArtifactCompileVersion() throws AzureExecutionException {
@@ -218,12 +218,13 @@ public class DeployMojo extends AbstractFunctionMojo {
         final FunctionAppConfig config = parser.parseConfig();
         final boolean newFunctionApp = !app.exists();
         final FunctionAppConfig defaultConfig = !newFunctionApp ?
-            fromFunctionApp(app, Objects.requireNonNull(app.getAppServicePlan())) :
-            buildDefaultConfig(config.subscriptionId(), config.resourceGroup(), config.appName());
+            fromFunctionApp(app) : buildDefaultConfig(config.subscriptionId(), config.resourceGroup(), config.appName());
         mergeAppServiceConfig(config, defaultConfig);
         if (!newFunctionApp && !config.disableAppInsights() && StringUtils.isEmpty(config.appInsightsKey())) {
             // fill ai key from existing app settings
-            config.appInsightsKey(Objects.requireNonNull(app.getAppSettings()).get(CreateOrUpdateFunctionAppTask.APPINSIGHTS_INSTRUMENTATION_KEY));
+            Optional.ofNullable(app.getAppSettings())
+                .map(map -> map.get(CreateOrUpdateFunctionAppTask.APPINSIGHTS_INSTRUMENTATION_KEY))
+                .ifPresent(config::appInsightsKey);
         }
         return new CreateOrUpdateFunctionAppTask(config).doExecute();
     }