+++ azure-functions-maven-plugin/CHANGELOG.md
@@ -1,6 +1,9 @@
 # Change Log
 All notable changes to the "Maven Plugin for Azure Function" will be documented in this file.
 - [Change Log](#change-log)
+  - [1.31.0](#1280)
+  - [1.30.0](#1280)
+  - [1.29.0](#1280)
   - [1.28.0](#1280)
   - [1.27.0](#1270)
   - [1.13.0](#1130)
@@ -23,6 +26,12 @@ All notable changes to the "Maven Plugin for Azure Function" will be documented 
   - [1.2.2](#122)
   - [1.2.1](#121)
   - [1.2.0](#120)
+
+## 1.31.0
+- Support function app hosted in Azure Container Apps Environment
+
+## 1.30.0
+- Migrate to use `stacks` API to get and validate function runtime stacks
 
 ## 1.29.0
 - Support `SqlTrigger` template in `mvn azure-functions:add`